refactor(customer): isolate getMetrics servuce imp  using hexagonal structure
- Introduced CustomerMetricsPort as application output port
- Moved JPA query logic to infrastructure adapter
- Decoupled service layer from Spring Data JPA
- saving metrics DTO construction in application boundary
