+++ rpm-customer-ms/src/main/java/com/seek/rpm/customer/exception/GlobalExceptionHandler.java
@@ -15,7 +15,7 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.servlet.NoHandlerFoundException;
 
-import com.seek.rpm.customer.dto.CustomResponse;
+import com.seek.rpm.customer.application.dto.CustomResponse;
 
 @RestControllerAdvice
 public class GlobalExceptionHandler {
