+++ rpm-customer-ms/src/main/java/com/seek/rpm/customer/infrastructure/persistence/jpa/CustomerJpaRepository.java
@@ -1,21 +1,21 @@
-package com.seek.rpm.customer.repository;
+package com.seek.rpm.customer.infrastructure.persistence.jpa;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 import com.seek.rpm.customer.domain.Customer;
-import com.seek.rpm.customer.dto.CustomerMetricsDTO;
+// import com.seek.rpm.customer.dto.CustomerMetricsDTO;
 
 @Repository
-public interface CustomerRepository extends JpaRepository<Customer, Long> {
+public interface CustomerJpaRepository extends JpaRepository<Customer, Long> {
 
     @Query("""
-                SELECT new com.seek.rpm.customer.dto.CustomerMetricsDTO(
+                SELECT
                     AVG(c.age),
                     STDDEV_POP(c.age)
-                )
                 FROM Customer c
             """)
-    CustomerMetricsDTO getCustomerMetrics();
+    Object[] getCustomerMetrics();
+
 }
