--- rpm-customer-ms/src/main/java/com/seek/rpm/customer/service/CustomerService.java
@@ -1,17 +1,0 @@
-package com.seek.rpm.customer.service;
-
-import java.util.List;
-
-import com.seek.rpm.customer.dto.CustomerCreateDTO;
-import com.seek.rpm.customer.dto.CustomerDTO;
-import com.seek.rpm.customer.dto.CustomerMetricsDTO;
-
-public interface CustomerService {
-
-    List<CustomerDTO> getAllCustomer();
-
-    CustomerDTO createCustomer(CustomerCreateDTO dto);
-
-    CustomerMetricsDTO getMetrics();
-
-}
